include pipe to calculate cumulative sum of resources
