+++ src/app/bucket/displayobjectives.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { DisplayobjectivesPipe } from './displayobjectives.pipe';
+
+describe('DisplayobjectivesPipe', () => {
+  it('create an instance', () => {
+    const pipe = new DisplayobjectivesPipe();
+    expect(pipe).toBeTruthy();
+  });
+});
